Use Java 5.0 foreach loops Various other quickfixes suggested by IntelliJ.
git-svn-id: https://findbugs.googlecode.com/svn/trunk@5172 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3

+++ findbugs/src/java/edu/umd/cs/findbugs/detect/FindPuzzlers.java
@@ -101,7 +101,7 @@ public class FindPuzzlers extends BytecodeScanningDetector {
 				=  stack.getStackItem(1).getConstant();
 			if (rightHandSide != null && rightHandSide instanceof Integer) {
 				constantArgumentToShift = true;
-				valueOfConstantArgumentToShift = ((Integer)rightHandSide).intValue();
+				valueOfConstantArgumentToShift = ((Integer) rightHandSide);
 				if (valueOfConstantArgumentToShift < 0 || valueOfConstantArgumentToShift >= 32)
 				 bugReporter.reportBug(new BugInstance(this, "ICAST_BAD_SHIFT_AMOUNT", NORMAL_PRIORITY)
 						.addClassAndMethod(this)
@@ -111,7 +111,7 @@ public class FindPuzzlers extends BytecodeScanningDetector {
 				}
 				if (leftHandSide != null 
 					&& leftHandSide instanceof Integer
-					&&  ((Integer)leftHandSide).intValue()
+					&& ((Integer) leftHandSide)
 						> 0) {
 				// boring; lie so other detectors won't get concerned
 				constantArgumentToShift = true;
@@ -129,7 +129,7 @@ public class FindPuzzlers extends BytecodeScanningDetector {
 			OpcodeStack.Item item = stack.getStackItem(0);
 			Object o = item.getConstant();
 			if (o != null && o instanceof Integer) {
-				int v =  ((Integer)o).intValue();
+				int v = (Integer) o;
 				if (v < 0 || v > 11)
 				 bugReporter.reportBug(new BugInstance(this, "DMI_BAD_MONTH", NORMAL_PRIORITY)
 						.addClassAndMethod(this)
@@ -153,7 +153,7 @@ public class FindPuzzlers extends BytecodeScanningDetector {
 			OpcodeStack.Item item = stack.getStackItem(1);
 			Object o = item.getConstant();
 			if (o != null && o instanceof Integer) {
-				int v =  ((Integer)o).intValue();
+				int v = (Integer) o;
 				if (v < 0 || v > 11)
 				 bugReporter.reportBug(new BugInstance(this, "DMI_BAD_MONTH", NORMAL_PRIORITY)
 						.addClassAndMethod(this)
